requirements-dev.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
static3==0.7.0
whitenoise==5.2.0
xlrd==2.0.1
